requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
boto3==1.37.35
botocore==1.37.35
discord.py==2.5.2
frozenlist==1.5.0
idna==3.10
jmespath==1.0.1
multidict==6.3.2
propcache==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
s3transfer==0.11.4
six==1.17.0
typing_extensions==4.13.1
urllib3==2.4.0
yarl==1.19.0
